requirements.txt
backports.entry-points-selectable==1.1.0
Bootstrap-Flask==1.2.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
distlib==0.3.2
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
filelock==3.0.12
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
geographiclib==1.52
geopy==2.2.0
google-auth==1.27.0
google-auth-oauthlib==0.4.2
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
platformdirs==2.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyodbc==4.0.32
python-dotenv==0.19.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.22
urllib3==1.26.6
virtualenv==20.7.0
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
